psc: added streaming of angle positions in deci-degrees + compensation for rotation, zero and direction

+++ emBODY/eBcode/arch-arm/embot/hw/embot_hw_tlv493d.h
@@ -41,7 +41,7 @@ namespace embot { namespace hw { namespace tlv493d {
     };
     
     
-    using Position = std::int16_t; // the measurement units is: TBD
+    using Position = std::int32_t; // the measurement units is: 0.01 degrees
             
     bool supported(embot::hw::TLV493D h);
